거미 보스 idle 행동 수정, 패턴 4 추가

+++ Assets/Scripts/Core/WeaponBehaviour.cs
@@ -67,7 +67,7 @@ public class WeaponBehaviour : MonoBehaviour
         weapons.Add(StaticData.inst.CreateWeapon(weaponID));
     }
 
-    public void GetDamaged(int damage)
+    public void GetDamaged(float damage)
     {
         /*Weapon.health -= damage;
         if (Weapon.health <= 0)
